mycourse/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.3
Django==5.1.2
django-ckeditor==6.7.1
django-excel-response2==3.0.6
django-js-asset==2.2.0
django-six==1.0.5
djangorestframework==3.15.2
drf-yasg==1.21.8
excel-base==1.0.4
inflection==0.5.1
isoweek==1.3.3
packaging==24.2
pillow==11.0.0
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
screen==1.0.1
six==1.16.0
sqlparse==0.5.1
TimeConvert==3.0.13
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
xlwt==1.3.0
